Move lookup to a separate file
